Setup admin's page

+++ user_account/admin/email/act_delete_tb_email_templates.php
@@ -2,6 +2,8 @@
 <?php
 $v_email_id = isset($_GET['id'])?$_GET['id']:'0';
 settype($v_email_id, 'int');
-$cls_tb_email_templates->delete(array('email_id' => $v_email_id));
+if($v_email_id>0)
+	$cls_tb_email_templates->delete(array('email_id' => $v_email_id));
+$_SESSION['ss_tb_email_templates_redirect'] = 1;
 redir(URL.$v_admin_key);
 ?>
